+++ mainproject/cluster/src/main/scala/com/raphtory/core/actors/partitionmanager/Archivist/PhantomConnector.scala
@@ -1,9 +1,11 @@
 package com.raphtory.core.actors.partitionmanager.Archivist
 import com.datastax.driver.core.SocketOptions
+import com.outworkers.phantom.builder.query.RootSelectBlock
 import com.outworkers.phantom.connectors.CassandraConnection
 import com.outworkers.phantom.database.{Database, DatabaseProvider}
 import com.outworkers.phantom.dsl._
 import com.raphtory.core.model.graphentities.Vertex
+import com.raphtory.core.utils.HistoryOrdering
 
 import scala.collection.mutable
 import scala.collection.mutable.TreeMap
@@ -34,20 +36,19 @@ class RaphtoryDatabase(override val connector: CassandraConnection) extends Data
 
 object RaphtoryDB extends RaphtoryDatabase(Connector.default)
 
-case class VertexHistoryPoint (id: Long, time: Long, value: Boolean)
+case class VertexHistoryPoint (id: Long, oldestPoint: Long, value: Boolean)
 
 abstract class VertexHistory extends Table[VertexHistory, VertexHistoryPoint] {
   object id extends LongColumn with PartitionKey
-  object time extends LongColumn with PartitionKey
-  object value extends BooleanColumn
+  object time extends LongColumn
+  object history extends MapColumn[Long,Boolean]
 
-  def save(vertex:Vertex) = {
-    val id = vertex.getId
-    val oldestPoint = vertex.oldestPoint.get
-    val history = createHistory(vertex.previousState)
-    val query = s"INSERT INTO raphtory.vertexHistory (id, oldestpoint, history) VALUES (${id},${oldestPoint},$history);"
-    println(query)
-    session.execute(query)
+  def saveNew(id:Long,oldestPoint:Long,history:mutable.TreeMap[Long, Boolean]) = {
+    session.execute(s"INSERT INTO raphtory.vertexHistory (id, oldestpoint, history) VALUES (${id},${oldestPoint},${createHistory(history)});")
+  }
+
+  def save(id:Long,history:mutable.TreeMap[Long,Boolean]) = {
+    session.execute(s"UPDATE raphtory.vertexHistory SET history = history + ${createHistory(history)} WHERE id = $id;")
   }
 
   private def createHistory(history: mutable.TreeMap[Long, Boolean]):String = {
@@ -59,9 +60,30 @@ abstract class VertexHistory extends Table[VertexHistory, VertexHistoryPoint] {
     s.dropRight(2) + "}"
   }
 
+  def allVertexHistory(id:Long) : Future[List[(Long, Map[Long, Boolean])]] = {
+    RaphtoryDB.vertexHistory.select(_.id,_.history).where(_.id eqs id).fetch()
+  }
 }
 
 
+/*
+ create keyspace raphtory with replication = {'class':'SimpleStrategy','replication_factor':1}; use raphtory;
+ CREATE TABLE raphtory.vertexhistory ( id bigint PRIMARY KEY, oldestPoint bigint, history map<bigint,boolean> );
+ CREATE INDEX history_key ON raphtory.vertexhistory ( KEYS (history) );
 
-//CREATE TABLE raphtory.vertexhistory ( id bigint PRIMARY KEY, oldestPoint bigint, history map<bigint,boolean> );
-//CREATE INDEX history_key ON raphtory.vertexhistory ( KEYS (history) );
+
+
+
+ CREATE TABLE vertexHistory (
+id bigint,
+time bigint,
+value boolean,
+PRIMARY KEY (id, time)
+)
+
+INSERT INTO raphtory.vertexHistory (id, time, value)
+  VALUES (1, 1,true)
+
+INSERT INTO raphtory.vertexHistory (id, time, value)
+  VALUES (1, 2,false)
+* */
